Implement creating a zendesk request and success view
