+++ app/styles/app/modules/buttons.scss
@@ -1,6 +1,6 @@
 %button {
   display: inline-block;
-  padding: 0.7em 0.8rem 0.6em;
+  padding: 0.7em 0.8rem;
   cursor: pointer;
   white-space: nowrap;
   border-radius: 2px;
