+++ config/environment.js
@@ -77,6 +77,11 @@ module.exports = function (environment) {
     emojiPrepend: '',
     statusPageStatusUrl: 'https://pnpcptp8xh9k.statuspage.io/api/v2/status.json',
 
+    zendesk: {
+      apiHost: 'https://travisci.zendesk.com',
+      createRequestEndpoint: '/api/v2/requests.json'
+    },
+
     moment: {
       includeTimezone: 'subset'
     }
